+++ app/src/lib/listener/icon.service.spec.ts
