+++ app/src/lib/listener/icon.service.ts
