+++ app/src/lib/singleton.ts
@@ -1,11 +1,11 @@
 import { app as electronApp } from 'electron';
-import { CipherService } from './ipc/cipher.service';
-import { IconService } from './ipc/icon.service';
-import { KeyService } from './ipc/key.service';
-import { App } from './lib/app';
-import { IpcService } from './lib/ipc.service';
-import { WindowManager } from './lib/window-manager';
-import { EmitterManager } from 'src/lib/emitter-manager';
+import { App } from './app';
+import { EmitterManager } from './emitter/emitter-manager';
+import { IpcService } from './ipc.service';
+import { CipherService } from './listener/cipher.service';
+import { IconService } from './listener/icon.service';
+import { KeyService } from './listener/key.service';
+import { WindowManager } from './window/window-manager';
 
 export const app = new App(electronApp);
 export const emitterManager: EmitterManager = new EmitterManager();
