+++ front/src/app/auth/login/login.component.ts
@@ -1,5 +1,5 @@
 import { AfterViewInit, Component, ElementRef } from '@angular/core';
-import { ElectronService } from 'src/app/base/electron/electron.service';
+import { ElectronService } from 'src/app/base/electron/emitter/electron.service';
 import { AuthenticationService } from 'src/app/base/services/authentication.service';
 import { WindowService } from 'src/app/base/services/window.service';
