+++ front/src/app/base/electron/emitter/cipher.service.ts
