+++ front/src/app/base/electron/emitter/electron.service.ts
