+++ front/src/app/base/electron/emitter/icon.service.ts
