+++ front/src/app/base/repositories/user-account.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { AbstractRepository } from './abstract.repository';
-import { KeyService } from 'src/app/base/electron/key.service';
+import { KeyService } from 'src/app/base/electron/emitter/key.service';
 import { UserAccount } from 'src/models';
 
 @Injectable({
