+++ front/src/app/main/components/account-editor/account-editor.component.ts
@@ -3,7 +3,7 @@ import { FormGroup, FormControl } from '@angular/forms';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { BehaviorSubject, combineLatest, from, Subscription } from 'rxjs';
 import { debounceTime, map, mergeMap } from 'rxjs/operators';
-import { IconService } from 'src/app/base/electron/icon.service';
+import { IconService } from 'src/app/base/electron/emitter/icon.service';
 import { OuterAccountRepository } from 'src/app/base/repositories/outer-account.repository';
 import { AutoUnsubscribe } from 'src/app/utils/autoUnsubscribe.decorator';
 import { genPassword } from 'src/app/utils/password';
