+++ front/src/app/main/components/sidebar/sidebar.component.ts
@@ -1,6 +1,6 @@
 import { AfterViewChecked, Component, ElementRef, Input, ViewChild } from '@angular/core';
 import { Observable } from 'rxjs';
-import { ElectronService } from 'src/app/base/electron/electron.service';
+import { ElectronService } from 'src/app/base/electron/emitter/electron.service';
 
 export type SidebarItem = {
   title: string;
