+++ front/src/app/main/components/tabbar/tabbar.component.ts
@@ -1,6 +1,6 @@
 import { CdkDragDrop, moveItemInArray } from '@angular/cdk/drag-drop';
 import { Component, EventEmitter, Output } from '@angular/core';
-import { ElectronService } from 'src/app/base/electron/electron.service';
+import { ElectronService } from 'src/app/base/electron/emitter/electron.service';
 import { AuthenticationService } from 'src/app/base/services/authentication.service';
 import { CsvService } from 'src/app/base/services/csv.service';
 import { TabService } from 'src/app/main/services/tab.service';
