+++ front/src/app/user-account-manager/user-account-manager.component.spec.ts
@@ -4,7 +4,7 @@ import { MatIconModule } from '@angular/material/icon';
 import { RouterTestingModule } from '@angular/router/testing';
 import { of } from 'rxjs';
 import { UserAccountManagerComponent } from './user-account-manager.component';
-import { ElectronService } from 'src/app/base/electron/electron.service';
+import { ElectronService } from 'src/app/base/electron/emitter/electron.service';
 
 import { UserAccountRepository } from 'src/app/base/repositories/user-account.repository';
 import { TestModule } from 'src/app/test/test.module';
